# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Agrafa backend API for agent ingestion, inventory creation, and read-side observability queries.",
        "title": "Agrafa API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/project-invitations": {
            "get": {
                "description": "Returns project invitations for the requested project.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project-invitations"
                ],
                "summary": "List project invitations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ProjectInvitationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a project invitation, stores a hashed invite token, and sends the invite email.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project-invitations"
                ],
                "summary": "Create project invitation",
                "parameters": [
                    {
                        "description": "Project invitation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ProjectInvitationCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ProjectInvitationCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/project-invitations/accept": {
            "post": {
                "description": "Accepts a valid invitation token for the authenticated user when the invite email matches the user email.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project-invitations"
                ],
                "summary": "Accept project invitation",
                "parameters": [
                    {
                        "description": "Project invitation acceptance payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ProjectInvitationAcceptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ProjectInvitationAcceptResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/project-invitations/by-token": {
            "get": {
                "description": "Returns safe invitation details when the invite token is valid and not expired or already accepted.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project-invitations"
                ],
                "summary": "Get project invitation by token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Raw invitation token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ProjectInvitationLookupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/project-invitations/{id}": {
            "delete": {
                "description": "Deletes a project invitation by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project-invitations"
                ],
                "summary": "Delete project invitation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project invitation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
...
        "types.ProjectInvitationAcceptRequest": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "raw-project-invitation-token"
                }
            }
        },
        "types.ProjectInvitationAcceptResponse": {
            "type": "object",
            "properties": {
                "already_member": {
                    "type": "boolean",
                    "example": false
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "types.ProjectInvitationCreateItemRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "teammate@example.com"
                },
                "role": {
                    "type": "string",
                    "example": "viewer"
                }
            }
        },
        "types.ProjectInvitationCreateRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "teammate@example.com"
                },
                "invitations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ProjectInvitationCreateItemRequest"
                    }
                },
                "project_id": {
                    "type": "integer",
                    "example": 1
                },
                "role": {
                    "type": "string",
                    "example": "viewer"
                }
            }
        },
        "types.ProjectInvitationCreateResponse": {
            "type": "object",
            "properties": {
                "project_id": {
                    "type": "integer",
                    "example": 1
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ProjectInvitationCreateResultDocument"
                    }
                }
            }
        },
        "types.ProjectInvitationCreateResultDocument": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "teammate@example.com"
                },
                "error_code": {
                    "type": "string",
                    "example": "already_invited"
                },
                "error_message": {
                    "type": "string",
                    "example": "An active invitation already exists for this email."
                },
                "invitation": {
                    "$ref": "#/definitions/types.ProjectInvitationDocument"
                },
                "role": {
                    "type": "string",
                    "example": "viewer"
                },
                "status": {
                    "type": "string",
                    "example": "created"
                }
            }
        },
        "types.ProjectInvitationDocument": {
            "type": "object",
            "properties": {
                "accepted_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "email": {
                    "type": "string",
                    "example": "teammate@example.com"
                },
                "expires_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "id": {
                    "type": "string",
                    "example": "pinv_0123456789abcdef"
                },
                "invited_by_user_id": {
                    "type": "string",
                    "example": "usr_0123456789abcdef"
                },
                "project_id": {
                    "type": "integer",
                    "example": 1
                },
                "role": {
                    "type": "string",
                    "example": "viewer"
                }
            }
        },
        "types.ProjectInvitationLookupDocument": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "teammate@example.com"
                },
                "expires_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "id": {
                    "type": "string",
                    "example": "pinv_0123456789abcdef"
                },
                "project_id": {
                    "type": "integer",
                    "example": 1
                },
                "project_name": {
                    "type": "string",
                    "example": "Agrafa Team"
                },
                "role": {
                    "type": "string",
                    "example": "viewer"
                }
            }
        },
        "types.ProjectInvitationLookupResponse": {
            "type": "object",
            "properties": {
                "project_invitation": {
                    "$ref": "#/definitions/types.ProjectInvitationLookupDocument"
                }
            }
        },
        "types.ProjectInvitationsResponse": {
            "type": "object",
            "properties": {
                "project_invitations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ProjectInvitationDocument"
                    }
                }
            }
        },
...
    }
}